--- project/css/style.css
@@ -1,68 +1,0 @@
-body {
-    background-color: #f8f9fa; /* Fundo claro e elegante */
-    font-family: 'Arial', sans-serif; /* Fonte agradável */
-}
-
-.form-login {
-    margin-top: 5vw;
-    background-color: #ffffff; /* Fundo branco para o formulário */
-    border-radius: 15px; /* Bordas arredondadas */
-    box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1); /* Sombra para destacar */
-    padding: 30px; /* Espaçamento interno */
-}
-
-h2 {
-    color: #343a40; /* Cor do título */
-}
-
-.form-label {
-    font-weight: bold;
-    color: #495057; /* Texto sutilmente escuro */
-}
-
-.form-control {
-    border: 2px solid #ced4da; /* Borda com melhor contraste */
-    padding: 10px;
-}
-
-.form-control:focus {
-    border-color: #28a745; /* Cor de foco verde */
-    box-shadow: 0 0 5px #28a745; /* Efeito de foco */
-}
-
-.btn-verde {
-    background-color: green !important; /* Cor verde definida */
-    border-color: green !important;    /* Borda verde */
-    color: white !important;           /* Texto branco */
-}
-
-.btn-verde:hover {
-    background-color: darkgreen !important; /* Alteração ao passar o mouse */
-}
-
-.row justify-content-center mb-10 {
-    margin-top: 5vw; /* Espaçamento superior */
-}
-.col-md-10 text-center {
-    margin-top: 2vw; /* Espaçamento superior */
-}
-
-.card {
-    background-color: #ffffff; /* Fundo branco */
-    border-radius: 15px; /* Bordas arredondadas */
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1); /* Sombra elegante */
-}
-
-.card h2 {
-    color: #343a40; /* Cor do título */
-}
-
-.card .form-label {
-    font-weight: bold;
-    color: #495057;
-}
-
-.card .form-control:focus {
-    border-color: #28a745; /* Foco verde */
-    box-shadow: 0 0 5px #28a745;
-}
