--- project/css/style.index.css
@@ -1,13 +1,0 @@
-body {
-    font-family: Arial, sans-serif;
-    background-color: #f5f5f5;
-}
-
-.card {
-    border: none;
-    transition: transform 0.3s ease;
-}
-
-.card:hover {
-    transform: scale(1.05); /* Efeito ao passar o mouse */
-}
